Add files via upload
The code tries to simulate different types of gates(AND, OR, NOT, ANDNOT, XOR).
XOR is not linearly separable and requires 2 layers of neurons and hence the input has been altered.
XOR takes 2 ANDNOT gate inputs that are sent through an OR gate so that they combine to form a representative gate.

The TC classifier is a basic representation of a 2-letter classifier represented on a 3x3 grid and can be extended to any size depending on the size of the array to be taken. It's specified to recognize the characters and give output 1 or 0 depending on the specification.